+++ lib/pepper/http/body_encoder/text.ex
@@ -1,0 +1,7 @@
+defmodule Pepper.HTTP.BodyEncoder.Text do
+  def encode_body(term, _options) do
+    blob = IO.iodata_to_binary(term)
+    headers = [{"content-type", "text/plain"}]
+    {:ok, {headers, blob}}
+  end
+end