+++ lib/pepper/http/response.ex
@@ -1,5 +1,6 @@
 defmodule Pepper.HTTP.Response do
   defstruct [
+    ref: nil,
     request: nil,
     headers: [],
     protocol: :unknown,
@@ -10,9 +11,11 @@ defmodule Pepper.HTTP.Response do
     body: "",
     data: nil,
     status_code: nil,
+    time: nil,
   ]
 
   @type t :: %__MODULE__{
+    ref: reference(),
     request: Pepper.HTTP.Request.t(),
     headers: [{String.t(), String.t()}],
     protocol: :unknown | :http1 | :http2,
@@ -23,5 +26,6 @@ defmodule Pepper.HTTP.Response do
     body: binary(),
     data: any(),
     status_code: non_neg_integer() | nil,
+    time: integer(),
   }
 end